+++ client/header/activity-panel/panels/inbox/index.js
@@ -5,7 +5,6 @@ import { __ } from '@wordpress/i18n';
 import { Component, Fragment } from '@wordpress/element';
 import { compose } from '@wordpress/compose';
 import { withDispatch } from '@wordpress/data';
-import { filter } from 'lodash';
 
 /**
  * Internal dependencies
@@ -17,6 +16,7 @@ import InboxNoteCard from './card';
 import { EmptyContent, Section } from '@woocommerce/components';
 import { QUERY_DEFAULTS } from 'wc-api/constants';
 import withSelect from 'wc-api/with-select';
+import { getUnreadNotesCount, hasValidNotes } from './utils';
 
 class InboxPanel extends Component {
 	constructor( props ) {
@@ -29,25 +29,6 @@ class InboxPanel extends Component {
 			activity_panel_inbox_last_read: this.mountTime,
 		};
 		this.props.updateCurrentUserData( userDataFields );
-	}
-
-	getUnreadNotesCount() {
-		const { lastRead, notes } = this.props;
-
-		const unreadNotes = filter( notes, ( note ) => {
-			const {
-				is_deleted: isDeleted,
-				date_created_gmt: dateCreatedGmt,
-			} = note;
-			if ( ! isDeleted ) {
-				return (
-					! lastRead ||
-					! dateCreatedGmt ||
-					new Date( dateCreatedGmt + 'Z' ).getTime() > lastRead
-				);
-			}
-		} );
-		return unreadNotes.length;
 	}
 
 	renderEmptyCard() {
@@ -64,16 +45,6 @@ class InboxPanel extends Component {
 				) }
 			</ActivityCard>
 		);
-	}
-
-	hasValidNotes() {
-		const { notes } = this.props;
-
-		const validNotes = filter( notes, ( note ) => {
-			const { is_deleted: isDeleted } = note;
-			return ! isDeleted;
-		} );
-		return validNotes.length > 0;
 	}
 
 	renderNotes( hasNotes ) {
@@ -95,7 +66,7 @@ class InboxPanel extends Component {
 	}
 
 	render() {
-		const { isError, isRequesting } = this.props;
+		const { isError, isRequesting, lastRead, notes } = this.props;
 
 		if ( isError ) {
 			const title = __(
@@ -120,7 +91,7 @@ class InboxPanel extends Component {
 			);
 		}
 
-		const hasNotes = this.hasValidNotes();
+		const hasNotes = hasValidNotes( notes );
 
 		return (
 			<Fragment>
@@ -131,7 +102,10 @@ class InboxPanel extends Component {
 							'Insights and growth tips for your business',
 							'woocommerce-admin'
 						) }
-						unreadMessages={ this.getUnreadNotesCount() }
+						unreadMessages={ getUnreadNotesCount(
+							notes,
+							lastRead
+						) }
 					/>
 				) }
 				<Section>
